claude-work/requirements.txt
# Core workflow
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# API integrations
tavily-python>=0.5.0
openai>=1.50.0

# Configuration & validation
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.0.0

# HTTP & async
httpx>=0.27.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Rate limiting & retry
tenacity>=8.2.0
aiolimiter>=1.1.0

# Image & audio processing
pillow>=10.0.0
pydub>=0.25.1

# Utilities
python-dateutil>=2.8.2
tiktoken>=0.5.0

# Development & testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
black>=24.0.0
ruff>=0.2.0
